refactor(console): improve component folder structure
