+++ gravitee-apim-console-webui/src/management/api/portal/plans/edit/1-general-step/plan-edit-general-step.component.scss
