+++ gravitee-apim-console-webui/src/management/api/portal/plans/edit/3-restriction-step/plan-edit-restriction-step.component.scss
