+++ gravitee-apim-console-webui/src/management/api/portal/plans/list/api-portal-plan-list.component.html
